% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDict.R
\name{DataDictionary}
\alias{DataDictionary}
\title{DataDictionary}
\description{
The class representing a canonical data dictionary
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataDictionary$new()}}
\item \href{#method-validate_datadict}{\code{DataDictionary$validate_datadict()}}
\item \href{#method-document_data}{\code{DataDictionary$document_data()}}
\item \href{#method-clone}{\code{DataDictionary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the DataDictionary with the data it documents and at least a mock data dictionary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDictionary$new(dataset_code, datadict)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_code}}{uniquely identifies a dataset}

\item{\code{datadict}}{an existing datadict to be validated. It is expected the user does
the initial work of at least a basic data dictionary.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_datadict"></a>}}
\if{latex}{\out{\hypertarget{method-validate_datadict}{}}}
\subsection{Method \code{validate_datadict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDictionary$validate_datadict(datadict)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-document_data"></a>}}
\if{latex}{\out{\hypertarget{method-document_data}{}}}
\subsection{Method \code{document_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDictionary$document_data(data, only_new = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDictionary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
